HSEARCH-3791 Move low-level Elasticsearch utils to the lowlevel packages
This will be more consistent with the Lucene backend.
