+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/lowlevel/index/mapping/impl/MetadataFields.java
@@ -4,15 +4,15 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.util.impl;
+package org.hibernate.search.backend.elasticsearch.lowlevel.index.mapping.impl;
 
-public class ElasticsearchFields {
+public final class MetadataFields {
 
 	private static final String INTERNAL_FIELD_PREFIX = "_";
 
 	private static final char PATH_SEPARATOR = '.';
 
-	private ElasticsearchFields() {
+	private MetadataFields() {
 	}
 
 	public static String internalFieldName(String fieldName) {
