+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/mapping/impl/DiscriminatorTypeNameMapping.java
@@ -19,7 +19,7 @@ import org.hibernate.search.backend.elasticsearch.logging.impl.Log;
 import org.hibernate.search.backend.elasticsearch.search.projection.impl.ProjectionExtractionHelper;
 import org.hibernate.search.backend.elasticsearch.search.projection.impl.SearchProjectionExtractContext;
 import org.hibernate.search.backend.elasticsearch.search.projection.impl.SearchProjectionRequestContext;
-import org.hibernate.search.backend.elasticsearch.util.impl.ElasticsearchFields;
+import org.hibernate.search.backend.elasticsearch.lowlevel.index.mapping.impl.MetadataFields;
 import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 
 import com.google.gson.JsonObject;
@@ -32,7 +32,7 @@ import com.google.gson.JsonPrimitive;
 public class DiscriminatorTypeNameMapping implements TypeNameMapping {
 	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
 
-	private static final String MAPPED_TYPE_FIELD_NAME = ElasticsearchFields.internalFieldName( "entity_type" );
+	private static final String MAPPED_TYPE_FIELD_NAME = MetadataFields.internalFieldName( "entity_type" );
 
 	private final TypeNameDiscriminatorSchemaRootContributor schemaRootContributor =
 			new TypeNameDiscriminatorSchemaRootContributor();
